backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
cryptography>=42.0.0
itsdangerous>=2.2.0
httpx>=0.27.0
anthropic>=0.40.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
apscheduler>=3.10.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
html2text>=2024.2.26
python-dateutil>=2.9.0
starlette>=0.40.0
